feat: app 디렉토리 구성 및 이관작업, Header 컴포넌트와 Admin / Cart 페이지 분리작업
