+++ src/basic/__tests__/origin.test.tsx
@@ -1,8 +1,9 @@
 // @ts-nocheck
-import { render, screen, fireEvent, within, waitFor } from '@testing-library/react';
+import { fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import { vi } from 'vitest';
-import App from '../App';
+
 import '../../setupTests';
+import { App } from '../app';
 
 describe('쇼핑몰 앱 통합 테스트', () => {
   beforeEach(() => {
