+++ src/basic/app/components/Header.tsx
@@ -1,0 +1,75 @@
+import type { CartItem } from "../../../types";
+
+type HeaderProps = {
+  isAdmin: boolean;
+  searchTerm: string;
+  setSearchTerm: (term: string) => void;
+  setIsAdmin: (isAdmin: boolean) => void;
+  cart: CartItem[];
+  totalItemCount: number;
+};
+
+export function Header({
+  cart,
+  isAdmin,
+  searchTerm,
+  setIsAdmin,
+  setSearchTerm,
+  totalItemCount
+}: HeaderProps) {
+  return (
+    <header className="sticky top-0 z-40 border-b bg-white shadow-sm">
+      <div className="mx-auto max-w-7xl px-4">
+        <div className="flex h-16 items-center justify-between">
+          <div className="flex flex-1 items-center">
+            <h1 className="text-xl font-semibold text-gray-800">SHOP</h1>
+            {/* 검색창 - 안티패턴: 검색 로직이 컴포넌트에 직접 포함 */}
+            {!isAdmin && (
+              <div className="ml-8 max-w-md flex-1">
+                <input
+                  type="text"
+                  value={searchTerm}
+                  onChange={(e) => setSearchTerm(e.target.value)}
+                  placeholder="상품 검색..."
+                  className="w-full rounded-lg border border-gray-300 px-4 py-2 focus:border-blue-500 focus:outline-none"
+                />
+              </div>
+            )}
+          </div>
+          <nav className="flex items-center space-x-4">
+            <button
+              onClick={() => setIsAdmin(!isAdmin)}
+              className={`rounded px-3 py-1.5 text-sm transition-colors ${
+                isAdmin ? "bg-gray-800 text-white" : "text-gray-600 hover:text-gray-900"
+              }`}
+            >
+              {isAdmin ? "쇼핑몰로 돌아가기" : "관리자 페이지로"}
+            </button>
+            {!isAdmin && (
+              <div className="relative">
+                <svg
+                  className="h-6 w-6 text-gray-700"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M3 3h2l.4 2M7 13h10l4-8H5.4M7 13L5.4 5M7 13l-2.293 2.293c-.63.63-.184 1.707.707 1.707H17m0 0a2 2 0 100 4 2 2 0 000-4zm-8 2a2 2 0 11-4 0 2 2 0 014 0z"
+                  />
+                </svg>
+                {cart.length > 0 && (
+                  <span className="absolute -right-2 -top-2 flex h-5 w-5 items-center justify-center rounded-full bg-red-500 text-xs text-white">
+                    {totalItemCount}
+                  </span>
+                )}
+              </div>
+            )}
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+}
