+++ src/basic/app/pages/index.ts
@@ -1,0 +1,2 @@
+export * from "./AdminPage";
+export * from "./CartPage";
